packages/database/gim_database/session.py
import os
from typing import AsyncGenerator

from sqlalchemy.ext.asyncio import create_async_engine
from sqlalchemy.orm import sessionmaker
from sqlmodel.ext.asyncio.session import AsyncSession
from dotenv import load_dotenv

project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '../../..'))
try:
    load_dotenv(os.path.join(project_root, '.env.local'))
except PermissionError:
    # Some environments may deny access to local secret files.
    # Configuration should fall back to process environment variables or .env.
    pass
load_dotenv(os.path.join(project_root, '.env'))

DATABASE_URL = os.getenv("DATABASE_URL", "")

# Convert to asyncpg driver format
if DATABASE_URL.startswith("postgresql://"):
    DATABASE_URL = DATABASE_URL.replace("postgresql://", "postgresql+asyncpg://")

engine = create_async_engine(
    DATABASE_URL,
    echo=False,
    pool_pre_ping=True,
    connect_args={
        # Required for transaction pooler compatibility.
        "prepared_statement_cache_size": 0,
        "statement_cache_size": 0,
    },
)

async_session_factory = sessionmaker(
    engine,
    class_=AsyncSession,
    expire_on_commit=False,
)


async def get_async_session() -> AsyncGenerator[AsyncSession, None]:
    async with async_session_factory() as session:
        yield session
