packages/database/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gim-database"
version = "0.1.0"
description = "Shared database models and migrations for IssueIndex"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["gim_database"]
